Vanilla_VAE/requirements.txt
torch
PyYAML
wandb
